requirements.txt
streamlit==0.63.0
pandas==1.0.5
numpy==1.19.5
matplotlib==2.2.2
beautifulsoup4==4.9.0
requests==2.24.0
networkx==2.4
pillow==7.2.0
